// frontend/src/generated/model/organization-member-dto.ts
/**
 * merged spec
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { OrganizationRoleEnum } from './organization-role-enum';


/**
 * Represents a member of an organization.
 */
export interface OrganizationMemberDto { 
    userId?: number;
    username: string;
    role: OrganizationRoleEnum;
    email: string;
}