// frontend/src/generated/model/organization-member-list-paginated-dto.ts
/**
 * merged spec
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { Pageable } from './pageable';
import { OrganizationMemberDto } from './organization-member-dto';
import { Sort } from './sort';


export interface OrganizationMemberListPaginatedDto { 
    totalElements: number;
    totalPages: number;
    sort: Sort;
    first: boolean;
    last: boolean;
    number: number;
    pageable: Pageable;
    numberOfElements: number;
    size: number;
    empty: boolean;
    content: Array<OrganizationMemberDto>;
}